Basic emulation
I implemented only one instruction, the way that it is written is not
very good, and can be improved upon significantly. I will be refining
how operations are handled. I am thinking of making a simple function
array that maps to each operation. Element zero doing no operation,
element one performing a move operation and so on. I would also like to
clean up the indexing of where the register is at, so that it can just
be used as a function, or some way where I do not have to use if
statements.