Memory update, CLI usage
I made it so that memory can now be dynamically allocated by giving a
parameter to the command.

I also made a emulator.c fo the definitions of emulator.h.

Lastly, I added a CLI Usage section to the readme, and when the command
is passed -h then the program will print a link to the section. This
section will describe to use the emulator, and how it can be configured.