Progress, IVT
I added in all the current possible operations and started working on an
IVT.

However, before writing the code I will be making some documentation
both as a plan for what each function does and how it works, as well as
that I won't have to make the documentation later.

I will be looking into better programs to do documentation.