Added operators
I added in the operator functions and implemented the int operator,
the call operator, the return operator, and the jump operator.

I also created a stack, and the call operator stores the instruction
pointer on the stack to jump back to when the return operator is called.

Note: The return operator jumps to the current value on the top of the
stack and has no way to determine if it is the right address.