Works
The jmp instruction and call instruction now have proper logic.
By this I mean that the place to jump to's address is decremented by 12,
as after each instruction the IP is always incremented by 12.

I also printed "Hello\n\n\r" to the screen, with a jmp statement at the
end making it a loop, interrupt_test.asm.