Implemented conditional jmps
I got simple logic working for the cmp operator. I need to implement
label checks, but so far it is working fine.
I implemented conditional jmps determined by flags. I can see some
errors that might happen around the less than equal flag, but I think
it is fine for now.
I also changed the JZ, JC, and their nots alternatives into the JG and
JL and their nots. It is yet to be implemented in the assembler.
The next commit of the assembler will synchronize the labels of
instructions and their operation codes.