Documented and fixed code
I added in some basic comments that dictate what certain lines or
functions do. For the operations I implemented two more cases, and have
commented in the MOV_OPERATOR function what the assembly code would look
like for each case.
As mentioned in the assembler, I need to add in sections such as data
and code, because the program recieved segmentation faults, which was
caused by the emulator trying to run the string "Hello World".
This is temporarily solved by `if (operation <= OPERATION_MAX)` in
main.c.