APPLY_OPERATION
The idea of this is to cut down on the amount of times the same thing is
repeated. For example, all operations up to XOR have the same structure
except for that the operation being applied is different.
Therefore for small changes, dozens of cases have to be changed.
The APPLY_OPERATION function intends to take some information
about the values, the registers, and the operation to be conducted, and
shortens down the dozens of cases down to only 9.