Interrupts
The clock interrupt takes away focus when it calls an interrupt.
It does this so often it acts sort of as a termination of the program.

There is also a weird issue with the compiler, the 0x30 byte again
with the ret and outb and outd not working with comments.