I am unable to clock
I cannot get the clock working reliably even though it works on ms.
I've made interrupt streaming so that interrupts don't get called over each
other rather if a interrupt is being worked on it is pushed to the back
and will be worked on when t he interrupts before it have finished. This
required the implementation of the iret operator.
However it still does not work, either the timer gets all the processing time
or the non interrupt code gets all the processing time.
Sometimes it is shared, but that causes weird issues.
I have no idea how to get this working. If I add a frame cap, the same
issues persist.