Clock works
I fixed the clock, as well as the back burner was not working as
intended. The issue with the clock was that the address that was
returned to would be incremented by 1 because there was
no IP_SET check. The backburner also gets filled with interrupts to do
and doesn't work with its current logic. I will work on fixing it.
Audio also works only about half the time which is odd, this means
that only starts every other time the program is started.