Strange error
On line 41 in fancy_expression.asm to line 43 is the way
to print a new line. For some reason this messes up the printing and
causes a string error to occur where already printed characters are
overriden with new ones.
I don't know how to explain this, the best way to show it is through
screenshots "strange_error_8_7_2022_a" and "strange_error_8_7_2022_b".